Monitor: Move Plot to own file
This makes sure that it and matplotlib is only imported when necessary.
